Split channel authz field policy
